Refactor API into separate module
Still some issues - direct calls to process.exit(), not returning
error codes - but a first step.
